+++ src/SpeedyWeather.jl
@@ -159,4 +159,5 @@ include("models/shallow_water.jl")
 include("models/primitive_dry.jl")
 include("models/primitive_wet.jl")
 include("models/tree.jl")
+include("models/set.jl")
 end